agregar fianza
crud de fianza editar fianza en contrato y eliminar

+++ app/Http/Controllers/AfianzadoraController.php
@@ -4,8 +4,10 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Models\Afianzadora;
+use App\Models\Fianza;
 use App\Models\User;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\DB;
 
 class AfianzadoraController extends Controller
 {
@@ -143,7 +145,48 @@ class AfianzadoraController extends Controller
      */
     public function destroy(Afianzadora $afianzadora)
     {
-        $afianzadora->delete();
-        return redirect()->route('afianzadoras.index');
+
+        //  return $afianzadora->id;
+
+
+         $busqueda=DB::table('fianzas')
+         ->join('contratos','fianzas.id_contrato','=','contratos.id')
+         ->join('afianzadoras','fianzas.id_afianzadora','=','afianzadoras.id')
+         ->where('contratos.estatus','=',0)
+         ->where('fianzas.id_afianzadora','=',$afianzadora->id)->count();
+
+        if($busqueda>0){
+
+            $mensaje_error='La afianzadora no se puede eliminar, cuenta con contratos activos';
+            return redirect()->route('afianzadoras.index')->with(compact('mensaje_error'));
+
+        }else{
+            $busqueda2=DB::table('fianzas')
+            ->join('contratos','fianzas.id_contrato','=','contratos.id')
+            ->join('afianzadoras','fianzas.id_afianzadora','=','afianzadoras.id')
+            ->select('fianzas.id')
+            ->where('contratos.estatus','=',1)
+            ->where('fianzas.id_afianzadora','=',$afianzadora->id)->get();
+
+
+
+            for($i=0; $i<count($busqueda2);$i++){
+                $emilinar=Fianza::where('id','=',$busqueda2[$i]->id)->delete();
+            }
+            
+              $afianzadora->delete();
+                return redirect()->route('afianzadoras.index');
+        }
+        
+
+
+        // return $contratoA;
+
+        // $fianza=Fianza::where('id_afianzadora','=',$afianzadora->id)->count();
+
+  
+
+        
+
     }
 }
